Address human lobby PR feedback
