+++ package.json
@@ -30,23 +30,23 @@
     "pw:install": "PLAYWRIGHT_BROWSERS_PATH=.playwright-browsers playwright install"
   },
   "dependencies": {
-    "@better-auth/prisma-adapter": "^1.6.9",
+    "@better-auth/prisma-adapter": "^1.6.10",
     "@paralleldrive/cuid2": "3.3.0",
     "@prisma/adapter-pg": "^7.8.0",
     "@prisma/client": "^7.8.0",
     "@socket.io/bun-engine": "^0.1.1",
-    "better-auth": "^1.6.9",
+    "better-auth": "^1.6.10",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dotenv": "^17.4.2",
     "lucide-react": "^1.14.0",
-    "next": "16.2.2",
-    "next-intl": "^4.11.0",
+    "next": "^16.2.6",
+    "next-intl": "^4.11.1",
     "pg": "^8.20.0",
     "prisma": "^7.8.0",
     "radix-ui": "^1.4.3",
-    "react": "^19.2.5",
-    "react-dom": "^19.2.5",
+    "react": "^19.2.6",
+    "react-dom": "^19.2.6",
     "server-only": "^0.0.1",
     "shadcn": "^4.7.0",
     "socket.io": "^4.8.3",
@@ -57,9 +57,9 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
-    "@tailwindcss/postcss": "^4.2.4",
+    "@tailwindcss/postcss": "^4.3.0",
     "@types/bun": "^1.3.13",
-    "@types/node": "^25.6.0",
+    "@types/node": "^25.6.2",
     "@types/pg": "8.20.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
@@ -72,7 +72,7 @@
     "stylelint-config-recess-order": "^7.7.0",
     "stylelint-config-standard": "40.0.0",
     "stylelint-order": "^8.1.1",
-    "tailwindcss": "^4.2.4",
+    "tailwindcss": "^4.3.0",
     "typescript": "^6.0.3"
   },
   "simple-git-hooks": {
